C test program to set the watering schedule
Program to remotely program the sprinkler controller with a copy of
the test script.

Commands are issued multiple of times until the command returns OK.
Timeout messages are displayed when the command fails to return a
response.

This command runs on Linux and assumes serial port /dev/ttyACM0
for the local LoRaSerial radio.

Conversion to Windows may be as simple as replacing the port name.