Sprinkler Server: Add weather station and sprinkler server support
A web-site consisting of several .PHP pages updates the sprinkler
controller configurations and their schedules.  As changes are made the
sprinkler server is notified via a write to a shared pipe.

The sprinkler server support includes:
* Loads the sprinkler controllers as they power on with their
configuration and schedule
* Getting notified of sprinkler controller database changes
* Reads the sprinkler controller database and determines differences
* Downloads the differences to the sprinkler controllers

The weather station support reads the rain guage and the wind speed

Sprinkler Server commands:
* ATI93 - Display the total wind gauge count
* ATI94 - Display wind counts
* ATI95 - Display wind gauge values
* ATI96 - Display the total rain gauge count
* ATI97 - Display rain counts
* ATI98 - Display rain gauge values

* ATI99 - Product Description